Improved folder structure and file naming - Seperated code base in Node_Editor/Framework into Core, CoreExtensions, Interface and SaveSystem subfolders - Put default code into Node_Editor/Default, including default calculation canvas&behaviour routines, graph canvas&behaviour example and runtime examples - Default nodes are still located under Node_Editor/Nodes
