+++ Node_Editor/Default/FloatCalculation.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d6d37d55559ce304fb864e95081c872e
+timeCreated: 1484482479
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
