+++ Node_Editor/Default/Graph Example.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b97dfd59d67e6784093a56608a38b5a4
+folderAsset: yes
+timeCreated: 1484482728
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
