+++ Node_Editor/Framework/Core.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: eaaf349b89a36ac418cdcf5ccad17b30
+folderAsset: yes
+timeCreated: 1484483054
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
