+++ Node_Editor/Framework/Core/Node.cs
