+++ Node_Editor/Framework/Core/NodeEditorState.cs
