+++ Node_Editor/Framework/Core/NodeGroup.cs
