+++ Node_Editor/Framework/Core/NodeKnob.cs.meta
