+++ Node_Editor/Framework/Core/NodeOutput.cs
