+++ Node_Editor/Framework/CoreExtensions.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 3dbc031d78034164eba81d46dd554b38
+folderAsset: yes
+timeCreated: 1484483100
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
