+++ Node_Editor/Framework/CoreExtensions/ConnectionTypes.cs
@@ -141,14 +141,4 @@ namespace NodeEditorFramework
 		string InKnobTex { get; }
 		string OutKnobTex { get; }
 	}
-
-	// TODO: Node Editor: Built-In Connection Types
-	public class FloatType : IConnectionTypeDeclaration 
-	{
-		public string Identifier { get { return "Float"; } }
-		public Type Type { get { return typeof(float); } }
-		public Color Color { get { return Color.cyan; } }
-		public string InKnobTex { get { return "Textures/In_Knob.png"; } }
-		public string OutKnobTex { get { return "Textures/Out_Knob.png"; } }
-	}
 }
