+++ Node_Editor/Framework/Interface.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 61bc799571b04954ea564aa72eacd528
+folderAsset: yes
+timeCreated: 1484483085
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
