+++ Node_Editor/Framework/Interface/NodeEditorCallbackReceiver.cs
