--- Node_Editor/Framework/NodeCanvasSceneSave.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine;
-using NodeEditorFramework;
-
-public partial class NodeCanvasSceneSave : MonoBehaviour 
-{
-	public string saveName;
-	public NodeCanvas savedNodeCanvas;
-}
