+++ Node_Editor/Framework/SaveSystem.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: e9bef099233dac444af4f7f522385beb
+folderAsset: yes
+timeCreated: 1484483188
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
