+++ Node_Editor/Framework/SaveSystem/NodeCanvasSceneSave.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+using NodeEditorFramework;
+
+namespace NodeEditorFramework
+{
+	public partial class NodeCanvasSceneSave : MonoBehaviour 
+	{
+		public string saveName;
+		public NodeCanvas savedNodeCanvas;
+	}
+}
